cleanup + add uuid Co-authored-by: Udai Arneja <udai.arneja11ua@gmail.com>
