+++ go.mod
@@ -2,4 +2,7 @@ module github.com/SOMAS2021/SOMAS2021
 
 go 1.17
 
-require github.com/divan/goabm v0.0.0-20190225193149-39b1c19c3a2d
+require (
+	github.com/divan/goabm v0.0.0-20190225193149-39b1c19c3a2d
+	github.com/google/uuid v1.3.0
+)
